+++ examples/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Small tests to adapt when developing your own experimental setup."""